requirements.txt
PySide6>=6.0
paho-mqtt>=1.6
pyqtgraph>=0.13
pandas>=1.5